Create BlackJackSim.py
This program simulates a game (shoe) of blackjack as you would play inside of a casino. You may also initialize as many players as needed who play basic strategy perfectly for european blackjack. This was a made as a proof of concept evaluator, as we intend to test out different betting and stategy methods to improve overall gain from Blackjack.